problem solved

+++ LeMansSounds Normal but Not work/app/src/main/java/com/example/lemanssounds/SuperBubble.java
@@ -65,7 +65,7 @@ public class SuperBubble extends Bubble {
         return listOfBubbles.get(0).getLonguitude();
     }
     @Override
-    public float getRadius()
+    public int getRadius()
     {
         return listOfBubbles.get(0).getRadius();
     }
